update: api通信
